Fix floating-point price rounding errors: Float → Numeric migration
Problem:
- All price fields used db.Float causing precision errors ($100 → $99.99)
- Frustrating user experience across entire platform

Solution:
- Changed all monetary fields to db.Numeric(10, 2) for exact decimal storage
- Created and ran database migration on production
- Updated 9 model files

Files modified:
- backend/app/models/package.py (price)
- backend/app/models/booking.py (amount, total_price)
- backend/app/models/campaign.py (proposed_price, budget)
- backend/app/models/collaboration.py (amount)
- backend/app/models/creator_profile.py (revision_fee, engagement_rate)
- backend/app/models/creator_subscription_plan.py (price)
- backend/app/models/subscription_plan.py (price_monthly, price_yearly, platform_fee_percentage)
- backend/app/models/subscription.py (last_payment_amount)
- backend/app/models/dispute.py (payout_percentage)

Migration:
- backend/migrations/fix_float_to_numeric_prices.py
- Converted 11 columns across 8 tables
- Ran successfully on production database

Impact:
- $100 stays $100 (no more $99.99)
- All prices now exact to 2 decimal places
- Percentages exact to 4 decimal places (engagement_rate)

+++ backend/app/models/booking.py
@@ -15,8 +15,8 @@ class Booking(db.Model):
     status = db.Column(db.String(20), default='pending')  # pending, accepted, in_progress, completed, cancelled, rejected
     booking_date = db.Column(db.DateTime, default=datetime.utcnow)
     completion_date = db.Column(db.DateTime)
-    amount = db.Column(db.Float, nullable=False)
-    total_price = db.Column(db.Float, nullable=False)  # Alias for amount (for compatibility)
+    amount = db.Column(db.Numeric(10, 2), nullable=False)
+    total_price = db.Column(db.Numeric(10, 2), nullable=False)  # Alias for amount (for compatibility)
     duration_days = db.Column(db.Integer, nullable=True)  # Duration from package or proposal
     payment_status = db.Column(db.String(20), default='pending')  # pending, paid, failed, refunded, verified
     payment_method = db.Column(db.String(20), nullable=True)  # paynow, bank_transfer (set when user chooses payment method)
